fix launcher

+++ python/tutorials/03-matrix-multiplication.py
@@ -161,14 +161,12 @@ import triton.language as tl
 #       provided configs
 
 
-'''
 @triton.autotune(
     configs=[
         triton.Config({'BLOCK_SIZE_M': 32, 'BLOCK_SIZE_N': 32, 'BLOCK_SIZE_K': 32, 'GROUP_SIZE_M': 4}, num_stages=2, num_warps=4),
     ],
     key=['M', 'N', 'K'],
 )
-'''
 @triton.jit
 def matmul_kernel(
     # Pointers to matrices
@@ -286,10 +284,10 @@ def matmul(a, b, activation=None):
         b.stride(0), b.stride(1),
         c.stride(0), c.stride(1),
         ACTIVATION=activation,
-        BLOCK_SIZE_M=32,
-        BLOCK_SIZE_N=32,
-        BLOCK_SIZE_K=32,
-        GROUP_SIZE_M=4,
+        #BLOCK_SIZE_M=32,
+        #BLOCK_SIZE_N=32,
+        #BLOCK_SIZE_K=32,
+        #GROUP_SIZE_M=4,
     )
     return c
 
@@ -310,88 +308,38 @@ b = torch.randn((1024, 1024), device='cuda', dtype=torch.float16)
 triton_output = matmul(a, b, activation=None)
 torch.cuda.synchronize()
 
+test_compile = False
 REPEAT = 100
-for i in range(REPEAT):
-    print(i)
-    shutil.rmtree('/home/chunwei/.triton/cache', ignore_errors=True)
 
-    start_time = time.time()
+if test_compile:
+    for i in range(REPEAT):
+        print(i)
+        shutil.rmtree('/home/chunwei/.triton/cache', ignore_errors=True)
 
-    #triton_output = matmul(a, b, activation=None)
-    triton.compile(matmul_kernel, **haha.compile_kwargs)
+        start_time = time.time()
 
-    #torch.cuda.synchronize()
+        #triton_output = matmul(a, b, activation=None)
+        triton.compile(matmul_kernel, **haha.compile_kwargs)
 
-    key = "total"
-    duration = time.time() - start_time
-    haha.dic[key] = haha.dic.get(key, 0) + duration
+        #torch.cuda.synchronize()
+
+        key = "total"
+        duration = time.time() - start_time
+        haha.dic[key] = haha.dic.get(key, 0) + duration
 
 
 # warmup
 triton_output = matmul(a, b, activation=None)
 
-REPEAT = 100
 start_time = time.time()
 for i in range(REPEAT):
-    triton.compile(matmul_kernel, **haha.compile_kwargs)
-    torch.cuda.synchronize()
-    key = "launch"
-    duration = time.time() - start_time
-    haha.dic[key] = haha.dic.get(key, 0) + duration
+    triton_output = matmul(a, b, activation=None)
+torch.cuda.synchronize()
+key = "launch"
+duration = time.time() - start_time
+haha.dic[key] = duration
 
 
 for key,value in haha.dic.items():
     print(key,value/REPEAT * 1e3)
 
-import sys
-sys.exit(0)
-
-torch_output = torch.matmul(a, b)
-print(f"triton_output={triton_output}")
-print(f"torch_output={torch_output}")
-if triton.testing.allclose(triton_output, torch_output):
-    print("✅ Triton and Torch match")
-else:
-    print("❌ Triton and Torch differ")
-
-
-
-# %%
-# Benchmark
-# --------------
-#
-# Square Matrix Performance
-# ~~~~~~~~~~~~~~~~~~~~~~~~~~
-# We can now compare the performance of our kernel against that of cuBLAS. Here we focus on square matrices, but feel free to arrange this script as you wish to benchmark any other matrix shape.
-
-
-@triton.testing.perf_report(
-    triton.testing.Benchmark(
-        x_names=['M', 'N', 'K'],  # argument names to use as an x-axis for the plot
-        x_vals=[
-            8192
-        ],  # different possible values for `x_name`
-        line_arg='provider',  # argument name whose value corresponds to a different line in the plot
-        # possible values for `line_arg``
-        line_vals=['cublas', 'triton'],
-        # label name for the lines
-        line_names=["cuBLAS", "Triton"],
-        # line styles
-        styles=[('green', '-'), ('green', '--'), ('blue', '-'), ('blue', '--')],
-        ylabel="TFLOPS",  # label name for the y-axis
-        plot_name="matmul-performance",  # name for the plot. Used also as a file name for saving the plot.
-        args={},
-    )
-)
-def benchmark(M, N, K, provider):
-    a = torch.randn((M, K), device='cuda', dtype=torch.float16)
-    b = torch.randn((K, N), device='cuda', dtype=torch.float16)
-    if provider == 'cublas':
-        ms, min_ms, max_ms = triton.testing.do_bench(lambda: torch.matmul(a, b), rep=100)
-    if provider == 'triton':
-        ms, min_ms, max_ms = triton.testing.do_bench(lambda: matmul(a, b), rep=100)
-    perf = lambda ms: 2 * M * N * K * 1e-12 / (ms * 1e-3)
-    return perf(ms), perf(max_ms), perf(min_ms)
-
-
-benchmark.run(show_plots=True, print_data=True)
